pyproject.toml
[tool.black]
target-version = ['py310']
include-trailing-comma = true
exclude = '''
/(
    \.eggs
  | \.git
  | \.venv
  | _build
  | build
  | dist
  | input-images
  | results
)/
'''

[tool.isort]
line_length = 88
force_grid_wrap = 0
combine_as_imports = true
skip = []
multi_line_output = 3
include_trailing_comma = true
lines_after_imports = 2
known_first_party = []
known_third_party = []

[tool.poetry]
name = "self-org-maps"
version = "0.1.0"
description = ""
authors = ["Ruben Cronie <rubencronie@gmail.com>"]
readme = "README.md"
packages = [{include = "som"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "10.0.1"
celery = {version = "5.2.7", extras = ["sqs"]}
fastapi = "0.79.0"
fastapi-utils = "0.2.1"
matplotlib = "^3.6.3"
numpy = "^1.22.4"
opencv-contrib-python = "4.5.5.64"
psutil = "^5.9.4"
pydantic = ">=1.10.0"
python-multipart = "^0.0.5"
pytransloadit = "0.2.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pre-commit = "^3.0.4"
isort = "5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
